Initial conversion to jinja templates
Initial conversion to jinja templates. The generated markdown is currently incomplete -
due to the nature of the data structures it looks like I'll need to add some user-defined
functions in order to populate the tables. (Maybe - I'm pretty sure jinja supports
Maps...)

I've reduced the number of databases being tested while working on this - it saves
a fair amount of time.

Added flyway initializer and test schema to docker instance... but currently only
for postgresql since it's unexpectedly loading the schemas for all databases.
(I'm sure I misunderstood something...) This was on my wishlist since it makes
it much easier to extend this application so it allows us to perform meaningful
performance comparisons after adding just a bit more code to run those tests.

I added that at this time because I also wanted to capture initial work on
creating an Entity-Relationship Diagram (ERD) via using jinja to create a
GraphViz .dot file. This may be split into a separate project in the future
but it is a natural fit when performing test-container based tests since it
allows us to see we're using the expected schema.