Expanded use of jinjava extensions + macros
Extended use of jinjava extensions - plus now use java reflection to
handle registering most of those extensions. The extensions are not
yet complete - some of the one-off properties listed in the summary
tables aren't handled yet.

Added two jinja macros (as supported by jinjava). They handle most,
not all, of the table generation so most things are now reduced to
a single line. (And a lot of frantic duck paddling under the water.)
That makes the templates cleaner and results much more consistent.